requirements.txt
streamlit==1.41.1
langchain>=0.2
langchain-openai>=0.1
python-dotenv>=1.0
